refactor atlasci into own submodule at
