+++ .github/workflows/ci.yml
@@ -71,10 +71,10 @@ jobs:
         working-directory: schema
       - name: Run cli tests
         run: go test -race ./...
-        working-directory: cmd
-      - name: Run ci tests
+        working-directory: cmd/atlascmd
+      - name: Run atlasci tests
         run: go test -race ./...
-        working-directory: internal/ci
+        working-directory: cmd/atlasci
 
   
   integration-mysql56:
